requirements (1).txt
streamlit==1.17.0
pandas==1.5.3
opencv-python==4.7.0.72
PyPDF2==3.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.3.0
numpy==1.23.5
openpyxl==3.0.10
